backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
pydantic==2.8.0
pydantic-settings==2.3.0
neo4j==5.14.0
python-dotenv==1.0.0
requests==2.31.0
GitPython==3.1.40
tree-sitter==0.20.4
anthropic==0.40.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
reportlab==4.0.7
aiohttp==3.9.1
